agent_system/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.31.0
httpx==0.26.0
python-multipart==0.0.6
psycopg2-binary==2.9.5  # Changed from ==3.1 to ==2.9.5 as suggested
faiss-cpu==1.7.4
numpy==1.22.4
# ollama-client==0.5.0
typing-extensions>=0.8.0  # Added for Python 3.10 compatibility
python-dotenv>=1.0.0      # Added for environment variables handling
